mongo-basic/src/main/java/com/lucas/mongobasic/product/ProductService.java
package com.lucas.mongobasic.product;

import lombok.RequiredArgsConstructor;
import lombok.extern.slf4j.Slf4j;
import org.springframework.stereotype.Service;
import org.springframework.transaction.annotation.Transactional;

import java.util.List;

@Service
@RequiredArgsConstructor
@Slf4j
public class ProductService {

    private final ProductRepository productRepository;

    @Transactional
    public Product saveProduct(Product product) {
        if(product.getId() != null) {
            return save(product);
        }else {
            return update(product);
        }
    }

    private Product save(Product product) {
        return productRepository.save(product);
    }

    private Product update(Product product) {
        return productRepository.save(product);
    }

    @Transactional(readOnly = true)
    public List<Product> findAll() {
        return productRepository.findAll();
    }
}
